pyproject.toml
[project]
name = "lekiwi-runtime"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.10"
dependencies = [
    "lerobot[feetech,pi]>=0.4.2",
    "livekit-agents[openai]>=1.1.7",
    "livekit-plugins-elevenlabs>=1.1.7",
    "livekit-plugins-noise-cancellation~=0.2",
    "livekit-plugins-silero>=1.1.7",
    "livekit>=1.0.20",
    "matplotlib>=3.10.7",
    "mediapipe>=0.10.14",
    "numpy>=2.2.6",
    "opencv-python>=4.12.0.88",
    "python-dotenv>=1.2.1",
    "pyzmq>=27.1.0",
    "scipy>=1.15.3",
    "sounddevice>=0.5.3",
    "timm>=1.0.22",
    "openai>=2.9.0",
]

[tool.uv]
package = true

[tool.uv.sources]
lerobot = { git = "https://github.com/huggingface/lerobot.git" }

[project.scripts]
lekiwi-host = "lekiwi.host:main"

[tool.setuptools]
packages = ["lekiwi", "nav"]
